Move base view model to abstract namespace
