+++ Signals/Signals/CoreLayer/Entities/StockItem.cs
@@ -39,6 +39,9 @@ public class StockItem : Entity
     public decimal LatestQuotedPrice { get; set; }
     public DateTime? WhenLatestQuoteReceived { get; set; }
 
+    public decimal AveragePurchasePrice { get; set; }
+    public DateTime WhenLastPurchased { get; set; }
+
 
     #region Non data members
