+++ Signals/Signals/Views/AddItemPageView.axaml
@@ -5,36 +5,66 @@
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              xmlns:vm="clr-namespace:Signals.ViewModels"
              xmlns:lang="clr-namespace:Signals.Resources"
-             x:DataType="vm:AddItemPageViewModel"   
+             x:DataType="vm:AddItemPageViewModel"
              x:Class="Signals.Views.AddItemPageView">
 
     <Design.DataContext>
         <vm:AddItemPageViewModel />
-    </Design.DataContext> 
+    </Design.DataContext>
 
     <!--
     Command="{Binding $parent[Window].((vm:MainViewModel)DataContext).GoToWatchlistDetailCommand, FallbackValue=null}"
      -->
-    
+
     <StackPanel>
         <StackPanel>
             <Label Content="{x:Static lang:Resources.SymbolLabel}"></Label>
-            <TextBox Text="{Binding Symbol, Converter={StaticResource UpperCaseConverter}}"></TextBox>
+            <TextBox Text="{Binding Symbol, 
+            Converter={StaticResource UpperCaseConverter}}"></TextBox>
         </StackPanel>
-        
-        <StackPanel Classes="formField" Orientation="Horizontal">
-            <CheckBox IsChecked="{Binding AddToWatchlist}"></CheckBox>
-            <Label Content="{x:Static lang:Resources.AddToWatchlistLabel}"></Label>
+
+        <StackPanel Classes="formField">
+            <CheckBox IsChecked="{Binding AddToWatchlist}"
+                      Content="{x:Static lang:Resources.AddToWatchlistLabel}">
+            </CheckBox>
         </StackPanel>
-        
-        <StackPanel Classes="formField" Orientation="Horizontal">
-            <CheckBox IsChecked="{Binding AddToHoldings}"></CheckBox>
-            <Label Content="{x:Static lang:Resources.AddToHoldingsLabel}"></Label>
+
+        <StackPanel Classes="formField">
+            <CheckBox IsChecked="{Binding AddToHoldings}"
+                      Content="{x:Static lang:Resources.AddToHoldingsLabel}">
+            </CheckBox>
+            <StackPanel IsVisible="{Binding  AddToHoldings}">
+                <Label Content="{x:Static lang:Resources.PriceLabel}"></Label>
+                <TextBox Name="PriceInput" Text="{Binding PurchasePrice }"
+                         IsEnabled="{Binding  AddToHoldings}" />
+                <Label Content="{x:Static lang:Resources.UnitsLabel}"></Label>
+                <TextBox Name="QuantityInput" Text="{Binding UnitsPurchased }"
+                         IsEnabled="{Binding  AddToHoldings}"
+                         HorizontalAlignment="Stretch" />
+                <Label Content="{x:Static lang:Resources.DateLabel}"></Label>
+                <TextBox Name="DateInput" Text="{Binding DatePurchased }"
+                         IsEnabled="{Binding  AddToHoldings}"
+                         HorizontalAlignment="Stretch" />
+                <StackPanel Orientation="Horizontal">
+                    <Label Content="{x:Static lang:Resources.HourLabel}"></Label>
+                    <Label>/</Label>
+                    <Label Content="{x:Static lang:Resources.MinuteLabel}"></Label>
+                </StackPanel>
+                <StackPanel Orientation="Horizontal" Spacing="5">
+                    <TextBox Name="HourInput" Text="{Binding HourPurchased }"
+                             IsEnabled="{Binding  AddToHoldings}"
+                             HorizontalAlignment="Stretch" />
+                    <TextBox Name="MinuteInput" Text="{Binding MinutePurchased }"
+                             IsEnabled="{Binding  AddToHoldings}"
+                             HorizontalAlignment="Stretch" />
+                </StackPanel>
+            </StackPanel>
         </StackPanel>
 
         <Button Content="Save" Command="{Binding AddSymbolCommand}"
                 CommandParameter="{Binding $parent[UserControl].((vm:AddItemPageViewModel)DataContext).BackLink, FallbackValue=null}"
-                IsEnabled="{Binding SaveIsEnabled}"></Button>
+                IsEnabled="{Binding SaveIsEnabled}">
+        </Button>
     </StackPanel>
-    
+
 </UserControl>
